refactor(project):清理整个项目，修复注释，删除过时代码，清理warning,修复大量大大小小的隐藏漏洞 fix(TVM):修复长整型未重载对应的乘方接口的问题 docs(Compiler):修正参数文档

+++ src/TVM/built_in/built_in_func.cpp
@@ -5,7 +5,6 @@
 #include <TVM/TRE.h>
 #include <TVM/TVM.h>
 #include <TVM/memory.h>
-#include <TVM/types/trc_string.h>
 #include <base/Error.h>
 #include <base/memory/objs_pool.hpp>
 #include <cstdio>
@@ -18,7 +17,7 @@ void LENGTH(int argc, TVM* vm) {
             error::ArgumentError, language::error::argumenterror, "len", "1");
     }
     auto firsts = (def::STRINGOBJ)vm->pop();
-    vm->push(new types::trc_int(firsts->len()));
+    vm->push(new types::trc_int((int)firsts->len()));
 }
 
 void HELP([[maybe_unused]] int argc, [[maybe_unused]] TVM* vm) {
